+++ src/detector.js
@@ -6,6 +6,9 @@ angular.module('angular-detector', [
 
 ]).provider('Detector', function() {
   var parser = new UAParser();
+  parser.newInstance = function(uastring, extension){
+  	return new UAParser(uastring, extension);
+  };
   parser['$get'] = function() {
     return parser;
   };